requirements.txt
python-dotenv
langchain
langchain-community 
langchain-openai
langchain-ollama
#pypdf
#faiss-cpu
chromadb
groq
langchain-groq
beautifulsoup4
streamlit
# langchain_google_genai
PyPDF2
langchain-chroma
#pdfplumber
memory_profiler
PyMuPDF
stqdm
huggingface-hub
#pysqlite3-binary 
gspread
oauth2client
pandas
google-api-python-client
pytesseract
google-generativeai
pillow
numpy
streamlit
opencv-python-headless
textblob
scipy
nltk
scikit-learn
pydantic
werkzeug
python-docx
openpyxl
nest_asyncio
toml
langchain-anthropic
anthropic
pdf2image
nest-asyncio
langchain_groq
PyJWT
plotly

# Add these CPU-only alternatives
tensorflow-cpu
tensorflow-text
tensorflow-hub
scikit-learn>=1.3.0
langchain-huggingface>=0.0.10